Do not create more than one PR with the same title

+++ vars/elifeGithubPullRequest.groovy
@@ -1,4 +1,9 @@
 def call(branch, title, description = '', base='master', temporaryFile = 'pull-request.log') {
+    def isThereAnExistingPullRequest = (sh script: "hub issue | grep \"$title\"", returnStatus: true) == 0
+    if (isThereAnExistingPullRequest) {
+        echo "There is already an existing PR with title: ${title}"
+        return
+    }
     sh "git push origin ${branch}"
     def fullText = title + "\n\n" + description
     writeFile(file: temporaryFile, text: fullText)
